refactor: Make core's API more clear (#492)
