--- core/src/sign/mod.rs
@@ -1,11 +1,0 @@
-mod api;
-pub use api::Build;
-pub use api::Key;
-pub use api::Load;
-
-mod request;
-pub use request::SigningMethod;
-pub use request::SigningRequest;
-
-mod signer;
-pub use signer::Signer;
