+++ core/src/signer.rs
